refactor: isolate alan shell command adapter
